WS-73:Toggle Map working
